refactor(shared): create shared directory and move some files
These are shared between config and engine and core.
