+++ src/shared/union-syntax.js
